test(app): run local file-path regression without bun
